Remove circular API reference
